pyproject.toml
[tool.poetry]
name = "testsite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
django = "4.2"
django-debug-toolbar = "4.2.0"
pillow = "10.2.0"
djangorestframework = "3.14.0"
drf-yasg = "1.21.7"
psycopg2 = "^2.9.9"
django-environ = "0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
